incorporate suggestions + put tests back #5338
