+++ mne/tests/test_morph.py
@@ -9,12 +9,8 @@ import pytest
 import numpy as np
 from mne import (SourceEstimate, VolSourceEstimate, VectorSourceEstimate,
                  read_evokeds, SourceMorph, read_source_morph,
-                 compute_morph_matrix, morph_data, read_source_estimate,
-                 grade_to_vertices, read_source_spaces)
-from numpy.testing import (assert_array_almost_equal, assert_array_equal,
-                           assert_allclose)
+                 read_source_estimate)
 from mne.datasets import sample, testing
-from mne.source_space import SourceSpaces
 from mne.minimum_norm import apply_inverse, read_inverse_operator
 from mne.utils import (run_tests_if_main, requires_nibabel, _TempDir,
                        requires_dipy, string_types)
@@ -30,8 +26,6 @@ subjects_dir = os.path.join(data_path, 'subjects')
 
 fname_evoked = os.path.join(sample_dir, 'sample_audvis-ave.fif')
 
-fname_fwd_vol = os.path.join(sample_dir,
-                             'sample_audvis_trunc-meg-vol-7-fwd.fif')
 fname_inv_vol = os.path.join(sample_dir,
                              'sample_audvis_trunc-meg-vol-7-meg-inv.fif')
 fname_vol = os.path.join(sample_dir,
@@ -40,7 +34,6 @@ fname_inv_surf = os.path.join(sample_dir,
                               'sample_audvis_trunc-meg-eeg-oct-6-meg-inv.fif')
 
 fname_smorph = os.path.join(sample_dir, 'sample_audvis_trunc-meg')
-fname_fmorph = os.path.join(sample_dir, 'fsaverage_audvis_trunc-meg')
 
 fname_t1 = os.path.join(subjects_dir, 'sample', 'mri', 'T1.mgz')
 
@@ -49,195 +42,6 @@ def _real_vec_stc():
     inv = read_inverse_operator(fname_inv_surf)
     evoked = read_evokeds(fname_evoked, baseline=(None, 0))[0].crop(0, 0.01)
     return apply_inverse(evoked, inv, pick_ori='vector')
-
-
-@requires_nibabel()
-@requires_dipy()
-@testing.requires_testing_data
-def test_save_vol_stc_as_nifti():
-    """Save the stc as a nifti file and export."""
-    import nibabel as nib
-    tempdir = _TempDir()
-    with warnings.catch_warnings(record=True):
-        warnings.simplefilter('always')
-        src = read_source_spaces(fname_fwd_vol)
-    vol_fname = os.path.join(tempdir, 'stc.nii.gz')
-
-    # now let's actually read a MNE-C processed file
-    stc = read_source_estimate(fname_vol, 'sample')
-    assert (isinstance(stc, VolSourceEstimate))
-
-    stc.save_as_volume(vol_fname, src,
-                       dest='surf', mri_resolution=False)
-    with warnings.catch_warnings(record=True):  # nib<->numpy
-        img = nib.load(vol_fname)
-    assert (img.shape == src[0]['shape'] + (len(stc.times),))
-
-    with warnings.catch_warnings(record=True):  # nib<->numpy
-        t1_img = nib.load(fname_t1)
-    stc.save_as_volume(os.path.join(tempdir, 'stc.nii.gz'), src,
-                       dest='mri', mri_resolution=True)
-    with warnings.catch_warnings(record=True):  # nib<->numpy
-        img = nib.load(vol_fname)
-    assert (img.shape == t1_img.shape + (len(stc.times),))
-    assert_allclose(img.affine, t1_img.affine, atol=1e-5)
-
-    # export without saving
-    img = stc.as_volume(src, dest='mri', mri_resolution=True)
-    assert (img.shape == t1_img.shape + (len(stc.times),))
-    assert_allclose(img.affine, t1_img.affine, atol=1e-5)
-
-    src = SourceSpaces([src[0], src[0]])
-    stc = VolSourceEstimate(np.r_[stc.data, stc.data],
-                            [stc.vertices, stc.vertices],
-                            tmin=stc.tmin, tstep=stc.tstep, subject='sample')
-    img = stc.as_volume(src, dest='mri', mri_resolution=False)
-    assert (img.shape == src[0]['shape'] + (len(stc.times),))
-
-    pytest.raises(ValueError, stc.as_volume, src, mri_resolution=(4., 4., 4.))
-
-
-@requires_nibabel()
-@requires_dipy()
-@pytest.mark.slowtest
-@testing.requires_testing_data
-def test_morph_data():
-    """Test morphing of data."""
-    subject_from = 'sample'
-    subject_to = 'fsaverage'
-
-    stc_from = read_source_estimate(fname_smorph, subject='sample')
-    stc_to = read_source_estimate(fname_fmorph)
-
-    # make sure we can specify grade
-    stc_from.crop(0.09, 0.1)  # for faster computation
-    stc_to.crop(0.09, 0.1)  # for faster computation
-    assert_array_equal(stc_to.time_as_index([0.09, 0.1], use_rounding=True),
-                       [0, len(stc_to.times) - 1])
-    with warnings.catch_warnings(record=True):
-        pytest.raises(ValueError, stc_from.morph, subject_to, grade=5,
-                      smooth=-1,
-                      subjects_dir=subjects_dir)
-
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter('always')
-        stc_to1 = stc_from.morph(subject_to, grade=3, smooth=12,
-                                 subjects_dir=subjects_dir)
-
-    stc_to1.save(os.path.join(tempdir, '%s_audvis-meg' % subject_to))
-    # Morphing to a density that is too high should raise an informative error
-    # (here we need to push to grade=6, but for some subjects even grade=5
-    # will break)
-    with warnings.catch_warnings(record=True):
-        pytest.raises(ValueError, stc_to1.morph, subject_from, grade=6,
-                      subjects_dir=subjects_dir)
-    # make sure we can specify vertices
-    vertices_to = grade_to_vertices(subject_to, grade=3,
-                                    subjects_dir=subjects_dir)
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter('always')
-        stc_to2 = morph_data(subject_from, subject_to, stc_from,
-                             grade=vertices_to, smooth=12,
-                             subjects_dir=subjects_dir)
-
-    # make sure we get a warning about # of steps
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter('always')
-        morph_data(subject_from, subject_to, stc_from,
-                   grade=vertices_to, smooth=1,
-                   subjects_dir=subjects_dir)
-    assert sum('consider increasing' in str(ww.message) for ww in w) == 2
-
-    assert_array_almost_equal(stc_to.data, stc_to1.data, 5)
-    assert_array_almost_equal(stc_to1.data, stc_to2.data)
-    # make sure precomputed morph matrices work
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter('always')
-        morph_mat = compute_morph_matrix(subject_from, subject_to,
-                                         stc_from.vertices, vertices_to,
-                                         smooth=12, subjects_dir=subjects_dir)
-    # compute_morph_matrix is deprecated
-    assert sum('deprecated' in str(ww.message) for ww in w) == 1
-
-    with warnings.catch_warnings(record=True):
-        stc_to2 = stc_from.morph_precomputed(subject_to, vertices_to,
-                                             morph_mat)
-    assert_array_almost_equal(stc_to1.data, stc_to2.data)
-
-    with warnings.catch_warnings(record=True):
-        pytest.raises(ValueError, stc_from.morph_precomputed,
-                      subject_to, vertices_to, 'foo')
-        pytest.raises(ValueError, stc_from.morph_precomputed,
-                      subject_to, [vertices_to[0]], morph_mat)
-        pytest.raises(ValueError, stc_from.morph_precomputed,
-                      subject_to, [vertices_to[0][:-1], vertices_to[1]],
-                      morph_mat)
-        pytest.raises(ValueError, stc_from.morph_precomputed, subject_to,
-                      vertices_to, morph_mat, subject_from='foo')
-
-    # steps warning
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter('always')
-        compute_morph_matrix(subject_from, subject_to,
-                             stc_from.vertices, vertices_to,
-                             smooth=1, subjects_dir=subjects_dir)
-
-    mean_from = stc_from.data.mean(axis=0)
-    mean_to = stc_to1.data.mean(axis=0)
-    assert (np.corrcoef(mean_to, mean_from).min() > 0.999)
-
-    # make sure we can fill by morphing
-    with warnings.catch_warnings(record=True):
-        stc_to5 = morph_data(subject_from, subject_to, stc_from, grade=None,
-                             smooth=12, subjects_dir=subjects_dir)
-    assert (stc_to5.data.shape[0] == 10242 + 10242)
-
-    # Morph sparse data
-    # Make a sparse stc
-    stc_from.vertices[0] = stc_from.vertices[0][[100, 500]]
-    stc_from.vertices[1] = stc_from.vertices[1][[200]]
-    stc_from._data = stc_from._data[:3]
-
-    with warnings.catch_warnings(record=True):
-        pytest.raises(RuntimeError, stc_from.morph, subject_to, sparse=True,
-                      grade=5, subjects_dir=subjects_dir)
-
-    with warnings.catch_warnings(record=True):
-        stc_to_sparse = stc_from.morph(subject_to, grade=None, sparse=True,
-                                       subjects_dir=subjects_dir)
-    assert_array_almost_equal(np.sort(stc_from.data.sum(axis=1)),
-                              np.sort(stc_to_sparse.data.sum(axis=1)))
-    assert len(stc_from.rh_vertno) == len(stc_to_sparse.rh_vertno)
-    assert len(stc_from.lh_vertno) == len(stc_to_sparse.lh_vertno)
-    assert stc_to_sparse.subject == subject_to
-    assert stc_from.tmin == stc_from.tmin
-    assert stc_from.tstep == stc_from.tstep
-
-    stc_from.vertices[0] = np.array([], dtype=np.int64)
-    stc_from._data = stc_from._data[:1]
-
-    with warnings.catch_warnings(record=True):
-        stc_to_sparse = stc_from.morph(subject_to, grade=None, sparse=True,
-                                       subjects_dir=subjects_dir)
-    assert_array_almost_equal(np.sort(stc_from.data.sum(axis=1)),
-                              np.sort(stc_to_sparse.data.sum(axis=1)))
-
-    assert len(stc_from.rh_vertno) == len(stc_to_sparse.rh_vertno)
-    assert len(stc_from.lh_vertno) == len(stc_to_sparse.lh_vertno)
-    assert stc_to_sparse.subject == subject_to
-    assert stc_from.tmin == stc_from.tmin
-    assert stc_from.tstep == stc_from.tstep
-
-    # Morph vector data
-    stc_vec = _real_vec_stc()
-
-    # Ignore warnings about number of steps
-    with warnings.catch_warnings(record=True) as w:
-        stc_vec_to1 = stc_vec.morph(subject_to, grade=3, smooth=12,
-                                    subjects_dir=subjects_dir)
-        stc_vec_to2 = stc_vec.morph_precomputed(subject_to, vertices_to,
-                                                morph_mat)
-    assert_array_almost_equal(stc_vec_to1.data, stc_vec_to2.data)
 
 
 @requires_nibabel()
